// src/app/Interfaces/expertise.ts
export interface expertiseContentTypes {
    slugName: string;
    mainHeadline: string;
    description: string;
    serviceHeading1: string;
    serviceDesc1: string;
    serviceHeading2: string;
    serviceDesc2: string;
    step1Heading: string;
    step1Desc: string;
    step2Heading: string;
    step2Desc: string;
    step3Heading: string;
    step3Desc: string;
    blackSecHeading: string;
    blackSecDesc: string;
    blackSecHeading1: string;
    blackSecDesc1: string;
    blackSecHeading2: string;
    blackSecDesc2: string;
    blackSecHeading3: string;
    blackSecDesc3: string;
    blackSecHeading4: string;
    blackSecDesc4: string;
    blackSecHeading5: string;
    blackSecDesc5: string;
  }
  